- Making fill_tables.php create forums, topics and posts. - Fixing forums up a bit. - Adding default LO display method.

+++ dokeos/repository/lib/learning_object/learning_path_chapter/learning_path_chapter.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `%prefix%learning_path_chapter` (
   `id` int(10) unsigned NOT NULL default '0',
-  `display_order` mediumint(8) unsigned NOT NULL default '0',
+  `display_order` int(10) unsigned NOT NULL default '0',
   PRIMARY KEY  (`id`)
 );
